requirements.txt
langchain
pypdf
sentence-transformers
openai
chromadb==0.3.29
langchain_community
streamlit_chat
python-dotenv
langchain_groq
langchain_experimental
